requirements.txt
newsapi-python
transformers
python-docx
diffusers 
scipy